+++ backend/backend/src/wrapper/entity/transaction/dto.rs
@@ -55,7 +55,7 @@ impl TransactionDTO {
         })
     }
 
-    pub(crate) async fn check_permissions(&self, user_id: i32) -> Result<bool, ApiError> {
+    pub(crate) async fn check_permissions(&self, user_id: i64) -> Result<bool, ApiError> {
         check_transaction_permissions(&self.budget_id, &self.source_id, &self.destination_id, user_id).await
     }
