+++ backend/backend/src/wrapper/permission/mod.rs
@@ -45,9 +45,9 @@ impl ToSchema<'static> for Permissions {
 
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
 pub(crate) struct PermissionsEntity {
-    pub(crate) user_id: i32,
+    pub(crate) user_id: i64,
     pub(crate) entity_type: String,
-    pub(crate) entity_id: i32,
+    pub(crate) entity_id: i64,
     pub(crate) permissions: Permissions,
 }
 
@@ -79,7 +79,7 @@ impl PermissionsEntity {
         Ok(false)
     }
 
-    pub(crate) async fn find_all_by_type_and_id(entity_type: &str, entity_id: i32) -> Result<Vec<Self>, ApiError> {
+    pub(crate) async fn find_all_by_type_and_id(entity_type: &str, entity_id: i64) -> Result<Vec<Self>, ApiError> {
         Ok(find_all(permissions::Entity::find_all_by_type_and_id(entity_type, entity_id))
             .await?
             .into_iter()
@@ -126,19 +126,19 @@ macro_rules! permission_impl {
 
 #[allow(unused)] //TODO remove once multiple users per account has been implemented
 pub(crate) trait Permission: PermissionByIds + WrapperEntity {
-    fn get_permissions(&self, user_id: i32) -> impl Future<Output = Result<Permissions, ApiError>> {
+    fn get_permissions(&self, user_id: i64) -> impl Future<Output = Result<Permissions, ApiError>> {
         async move { Self::get_permissions_by_id(self.get_id(), user_id).await }
     }
 
-    fn has_permission(&self, user_id: i32, permissions: Permissions) -> impl Future<Output = Result<bool, ApiError>> {
+    fn has_permission(&self, user_id: i64, permissions: Permissions) -> impl Future<Output = Result<bool, ApiError>> {
         async move { Self::has_permission_by_id(self.get_id(), user_id, permissions).await }
     }
 
-    fn add_permission(&self, user_id: i32, permissions: Permissions) -> impl Future<Output = Result<(), ApiError>> {
+    fn add_permission(&self, user_id: i64, permissions: Permissions) -> impl Future<Output = Result<(), ApiError>> {
         async move { Self::add_permission_by_id(self.get_id(), user_id, permissions).await }
     }
 
-    fn remove_permission(&self, user_id: i32, permissions: Permissions) -> impl Future<Output = Result<(), ApiError>> {
+    fn remove_permission(&self, user_id: i64, permissions: Permissions) -> impl Future<Output = Result<(), ApiError>> {
         async move { Self::remove_permission_by_id(self.get_id(), user_id, permissions).await }
     }
 }
@@ -146,7 +146,7 @@ pub(crate) trait Permission: PermissionByIds + WrapperEntity {
 pub(crate) trait HasPermissionOrError: Permission {
     fn has_permission_or_error(
         &self,
-        user_id: i32,
+        user_id: i64,
         permissions: Permissions,
     ) -> impl Future<Output = Result<(), ApiError>> {
         async move {
@@ -159,7 +159,7 @@ pub(crate) trait HasPermissionOrError: Permission {
 
 #[allow(unused)]
 pub(crate) trait PermissionByIds: TableName {
-    fn get_permissions_by_id(entity_id: i32, user_id: i32) -> impl Future<Output = Result<Permissions, ApiError>> {
+    fn get_permissions_by_id(entity_id: i64, user_id: i64) -> impl Future<Output = Result<Permissions, ApiError>> {
         async move {
             Ok(Permissions::from(
                 find_one(permissions::Entity::find_permission(user_id, Self::table_name(), entity_id)).await?,
@@ -168,8 +168,8 @@ pub(crate) trait PermissionByIds: TableName {
     }
 
     fn has_permission_by_id(
-        entity_id: i32,
-        user_id: i32,
+        entity_id: i64,
+        user_id: i64,
         permissions: Permissions,
     ) -> impl Future<Output = Result<bool, ApiError>> {
         async move {
@@ -181,8 +181,8 @@ pub(crate) trait PermissionByIds: TableName {
     }
 
     fn add_permission_by_id(
-        entity_id: i32,
-        user_id: i32,
+        entity_id: i64,
+        user_id: i64,
         permissions: Permissions,
     ) -> impl Future<Output = Result<(), ApiError>> {
         async move {
@@ -194,8 +194,8 @@ pub(crate) trait PermissionByIds: TableName {
     }
 
     fn remove_permission_by_id(
-        entity_id: i32,
-        user_id: i32,
+        entity_id: i64,
+        user_id: i64,
         permissions: Permissions,
     ) -> impl Future<Output = Result<(), ApiError>> {
         async move {
@@ -209,8 +209,8 @@ pub(crate) trait PermissionByIds: TableName {
 
 pub(crate) trait HasPermissionByIdOrError: PermissionByIds {
     fn has_permission_by_id_or_error(
-        entity_id: i32,
-        user_id: i32,
+        entity_id: i64,
+        user_id: i64,
         permissions: Permissions,
     ) -> impl Future<Output = Result<(), ApiError>> {
         async move {
@@ -237,9 +237,9 @@ fn has_permission_or_error_raw(
 }
 
 async fn save_permission_active_model(
-    entity_id: i32,
+    entity_id: i64,
     table_name: &str,
-    user_id: i32,
+    user_id: i64,
     permissions: Permissions,
 ) -> Result<(), ApiError> {
     let active_model = permissions::ActiveModel {
